requirements.txt
boto==2.49.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-storages-redux==1.3.3
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
psycopg2==2.8.2
python3-openid==3.1.0
pytz==2019.1
requests==2.21.0
requests-oauthlib==1.2.0
urllib3==1.24.1
wheel==0.24.0 
